[BUGFIX] Keep trying resource saves until successful
Previously, when we failed to successfully save a CachedResource object,
we would time out after 1 minute and then try again -- but if THAT
request failed, we would give up. Now we keep trying, every minute,
until the browser window is no longer open.

Maybe eventually we could add an exponential backoff, a'la gmail.